+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Liquid network: support for watch-only wallets
 - Watch-only wallets: support to delete credentials
 - Wallet onboard: automatic unique name creation support
+- Transaction summary dialog for hardware wallets
  
 ### Changed
 - Improve performance of login with hardware wallets